move cleanfilename
